requirements.txt
# Core dependencies
pandas>=1.3.0
pydantic>=1.8.2
python-Levenshtein>=0.12.2
python-dotenv>=0.19.0

# OpenAI dependencies
openai>=1.0.0

# Agent SDK dependencies
# Assuming 'agents' and 'ai' are internal or provided elsewhere,
# otherwise, specify installable package names/versions here.
# agents>=0.1.0
# ai>=0.1.0

# Testing dependencies
pytest>=6.2.5
pytest-mock>=3.6.1 